+++ src/test/java/com/example/ssauc/repositoryTests/BidRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.bid.entity.Bid;
 import com.example.ssauc.user.bid.repository.BidRepository;
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import static org.assertj.core.api.Assertions.assertThat;
-import java.math.BigDecimal;
+
 import java.time.LocalDateTime;
 
 @DataJpaTest
