+++ src/test/java/com/example/ssauc/repositoryTests/ChargeRepositoryTest.java
@@ -1,9 +1,9 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.login.entity.Users;
-import com.example.ssauc.user.mypage.entity.ReputationHistory;
+import com.example.ssauc.user.mypage.entity.Charge;
 import com.example.ssauc.user.login.repository.UsersRepository;
-import com.example.ssauc.user.mypage.repository.ReputationHistoryRepository;
+import com.example.ssauc.user.mypage.repository.ChargeRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -15,29 +15,32 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class ReputationHistoryRepositoryTest {
+public class ChargeRepositoryTest {
 
     @Autowired
-    private ReputationHistoryRepository reputationHistoryRepository;
+    private ChargeRepository chargeRepository;
 
     @Autowired
     private UsersRepository usersRepository;
 
     @Test
-    void testSaveReputationHistory() {
-        Users user = createTestUser("repUser", "repUser@example.com");
+    void testSaveCharge() {
+        Users user = createTestUser("chargeUser", "chargeUser@example.com");
         Users savedUser = usersRepository.save(user);
 
-        ReputationHistory history = ReputationHistory.builder()
+        Charge charge = Charge.builder()
                 .user(savedUser)
-                .changeType("Increase")
-                .changeAmount(0.5)
-                .newScore(5.0)
+                .impUid("imp_test")
+                .chargeType("Credit")
+                .amount(5000L)
+                .status("Completed")
+                .details("Charge details")
+                .receiptUrl("http://example.com/receipt")
                 .createdAt(LocalDateTime.now())
                 .build();
 
-        ReputationHistory savedHistory = reputationHistoryRepository.save(history);
-        assertThat(savedHistory.getHistoryId()).isNotNull();
+        Charge savedCharge = chargeRepository.save(charge);
+        assertThat(savedCharge.getChargeId()).isNotNull();
     }
 
     private Users createTestUser(String name, String email) {
@@ -47,7 +50,7 @@ public class ReputationHistoryRepositoryTest {
                 .password("password")
                 .phone("010-0000-0000")
                 .profileImage("http://example.com/img.png")
-                .reputation(4.3)
+                .reputation(4.2)
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
                 .lastLogin(LocalDateTime.now())
