+++ src/test/java/com/example/ssauc/repositoryTests/ChatMessageRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 
 import com.example.ssauc.user.chat.entity.ChatMessage;
