+++ src/test/java/com/example/ssauc/repositoryTests/ChatParticipantRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 
 import com.example.ssauc.user.chat.entity.ChatParticipant;
