+++ src/test/java/com/example/ssauc/repositoryTests/OrdersRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.order.entity.Orders;
 import com.example.ssauc.user.login.entity.Users;
