+++ src/test/java/com/example/ssauc/repositoryTests/ReportRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.chat.entity.Report;
 import com.example.ssauc.user.chat.repository.ReportRepository;
