+++ src/test/java/com/example/ssauc/repositoryTests/ReputationHistoryRepositoryTest.java
@@ -1,9 +1,9 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.login.entity.Users;
-import com.example.ssauc.user.mypage.entity.Withdraw;
+import com.example.ssauc.user.mypage.entity.ReputationHistory;
 import com.example.ssauc.user.login.repository.UsersRepository;
-import com.example.ssauc.user.mypage.repository.WithdrawRepository;
+import com.example.ssauc.user.mypage.repository.ReputationHistoryRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -15,31 +15,29 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class WithdrawRepositoryTest {
+public class ReputationHistoryRepositoryTest {
 
     @Autowired
-    private WithdrawRepository withdrawRepository;
+    private ReputationHistoryRepository reputationHistoryRepository;
 
     @Autowired
     private UsersRepository usersRepository;
 
     @Test
-    void testSaveWithdraw() {
-        Users user = createTestUser("withdrawUser", "withdrawUser@example.com");
+    void testSaveReputationHistory() {
+        Users user = createTestUser("repUser", "repUser@example.com");
         Users savedUser = usersRepository.save(user);
 
-        Withdraw withdraw = Withdraw.builder()
+        ReputationHistory history = ReputationHistory.builder()
                 .user(savedUser)
-                .amount(1000L)
-                .commission(50L)
-                .bank("Test Bank")
-                .account("1234567890")
-                .requestedAt(LocalDateTime.now())
-                .withdrawAt(LocalDateTime.now())
+                .changeType("Increase")
+                .changeAmount(0.5)
+                .newScore(5.0)
+                .createdAt(LocalDateTime.now())
                 .build();
 
-        Withdraw savedWithdraw = withdrawRepository.save(withdraw);
-        assertThat(savedWithdraw.getWithdrawId()).isNotNull();
+        ReputationHistory savedHistory = reputationHistoryRepository.save(history);
+        assertThat(savedHistory.getHistoryId()).isNotNull();
     }
 
     private Users createTestUser(String name, String email) {
@@ -49,7 +47,7 @@ public class WithdrawRepositoryTest {
                 .password("password")
                 .phone("010-0000-0000")
                 .profileImage("http://example.com/img.png")
-                .reputation(4.0)
+                .reputation(4.3)
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
                 .lastLogin(LocalDateTime.now())
