+++ src/test/java/com/example/ssauc/repositoryTests/SsaucApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
