+++ src/test/java/com/example/ssauc/repositoryTests/UserActivityRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.login.entity.Users;
 import com.example.ssauc.user.mypage.entity.UserActivity;
