+++ src/test/java/com/example/ssauc/repositoryTests/WithdrawRepositoryTest.java
@@ -1,9 +1,9 @@
-package com.example.ssauc;
+package com.example.ssauc.repositoryTests;
 
 import com.example.ssauc.user.login.entity.Users;
-import com.example.ssauc.user.mypage.entity.Charge;
+import com.example.ssauc.user.mypage.entity.Withdraw;
 import com.example.ssauc.user.login.repository.UsersRepository;
-import com.example.ssauc.user.mypage.repository.ChargeRepository;
+import com.example.ssauc.user.mypage.repository.WithdrawRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
@@ -15,32 +15,31 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 @DataJpaTest
 @AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
-public class ChargeRepositoryTest {
+public class WithdrawRepositoryTest {
 
     @Autowired
-    private ChargeRepository chargeRepository;
+    private WithdrawRepository withdrawRepository;
 
     @Autowired
     private UsersRepository usersRepository;
 
     @Test
-    void testSaveCharge() {
-        Users user = createTestUser("chargeUser", "chargeUser@example.com");
+    void testSaveWithdraw() {
+        Users user = createTestUser("withdrawUser", "withdrawUser@example.com");
         Users savedUser = usersRepository.save(user);
 
-        Charge charge = Charge.builder()
+        Withdraw withdraw = Withdraw.builder()
                 .user(savedUser)
-                .impUid("imp_test")
-                .chargeType("Credit")
-                .amount(5000L)
-                .status("Completed")
-                .details("Charge details")
-                .receiptUrl("http://example.com/receipt")
-                .createdAt(LocalDateTime.now())
+                .amount(1000L)
+                .commission(50L)
+                .bank("Test Bank")
+                .account("1234567890")
+                .requestedAt(LocalDateTime.now())
+                .withdrawAt(LocalDateTime.now())
                 .build();
 
-        Charge savedCharge = chargeRepository.save(charge);
-        assertThat(savedCharge.getChargeId()).isNotNull();
+        Withdraw savedWithdraw = withdrawRepository.save(withdraw);
+        assertThat(savedWithdraw.getWithdrawId()).isNotNull();
     }
 
     private Users createTestUser(String name, String email) {
@@ -50,7 +49,7 @@ public class ChargeRepositoryTest {
                 .password("password")
                 .phone("010-0000-0000")
                 .profileImage("http://example.com/img.png")
-                .reputation(4.2)
+                .reputation(4.0)
                 .createdAt(LocalDateTime.now())
                 .updatedAt(LocalDateTime.now())
                 .lastLogin(LocalDateTime.now())
